Refactor(DK-589): BookList 구조 수정
- BookListLayout과 BookList를 분리하여 outlet을 통해 렌더링하던 구조를 단일 BookList(index.tsx)에서 직접 불러오도록 변경
- 불필요한 중첩 구조를 제거하여 코드 간결화
