--- src/pages/Home/_home.module.scss
@@ -1,17 +1,0 @@
-@use "/src/styles/abstracts/" as *;
-
-.container {
-  @include respond-to(lg) {
-    padding: 1.6rem 16rem 1.6rem;
-  }
-
-  @include respond-to(md) {
-    padding: 1.6rem 4rem 1.6rem;
-  }
-
-  @include respond-to(sm) {
-    padding: 1.6rem 2rem 1.6rem;
-    flex-direction: column;
-    align-items: center;
-  }
-}
